requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Injector==0.12.3
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-pymysql==0.2.3
python-dotenv==0.14.0
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
pyodbc==4.0.30
PyJWT==1.7.1
requests==2.23.0
requests-toolbelt==0.9.1
importlib-metadata==1.6.0
multidict==4.7.6
PyJWT==1.7.1
